Grammatical improvements
Found this repository after browsing the website. The idea of the website was nice (having a open organization / community / etc), but the spelling and overall writing on the website was terrible, something that really bothers me. I tried to quickly fix some of the mistakes by giving it a proof-read. I hope you guys are not offended by this, I'm just reaching out to give a helping hand. Good luck!